+++ src/java/com/leafchild/cashmashine/entity/Customer.java
@@ -18,7 +18,7 @@ public class Customer implements Serializable {
 
     @Id
     @GeneratedValue
-    private Integer customer_id;
+    private Long customer_id;
     @NotNull
     private String customerName;
     @OneToMany
@@ -35,12 +35,12 @@ public class Customer implements Serializable {
         this.customerName = customerName;
     }
 
-    public Integer getCustomer_id(){
+    public Long getCustomer_id(){
 
         return customer_id;
     }
 
-    public void setCustomer_id(Integer customer_id){
+    public void setCustomer_id(Long customer_id){
 
         this.customer_id = customer_id;
     }
